refactory: move enums into module due to length of implementations
