+++ src/enums/mod.rs
@@ -1,0 +1,11 @@
+//! Module with the enums and their implementations.
+
+pub mod base_stat_enum;
+pub mod stat_enum;
+pub mod stat_modifier_enum;
+pub mod type_category_enum;
+
+pub use base_stat_enum::BaseStat;
+pub use stat_enum::Stat;
+pub use stat_modifier_enum::StatModifier;
+pub use type_category_enum::TypeCategory;
