+++ src/lib.rs
@@ -1,18 +1,18 @@
-//! Hello
+//! This library provides a set of generic components that could be used in a game.
 //!
 
 pub mod decimal_attribute;
+pub mod enums;
 pub mod errors;
 pub mod integer_attribute;
-pub mod stat_enum;
 pub mod statistic;
 pub mod traits;
 
 pub use decimal_attribute::DecimalAttribute;
+pub use enums::BaseStat;
+pub use enums::Stat;
+pub use enums::StatModifier;
+pub use enums::TypeCategory;
 pub use errors::AttributeError;
 pub use integer_attribute::IntegerAttribute;
-pub use stat_enum::BaseStat;
-pub use stat_enum::Stat;
-pub use stat_enum::StatModifier;
-pub use stat_enum::TypeCategory;
 pub use statistic::Statistic;
